% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{hud_colours}
\alias{hud_colours}
\title{HUD Colours}
\arguments{
\item{medium}{"print" (white background theme) or "web" (dark background theme)}

\item{colours}{Required number of colours}
}
\description{
ggplot2 colour scale generation function for HUD styles.
}
\examples{
hud_colours(medium = "web", colours = 2)
}
\keyword{colours}
\keyword{ggplot2}
\keyword{hud}
