% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hud_theme}
\alias{hud_theme}
\title{HUD Theme}
\usage{
hud_theme(medium = "web", layout = "big")
}
\arguments{
\item{medium}{"web" (dark background theme) or "print" (white background theme)}

\item{layout}{"big" or "small}
}
\description{
ggplot2 theme generator using HUD styles.
}
\examples{
hud_theme(medium = "web", layout = "big")
}
\keyword{ggplot2}
\keyword{hud}
\keyword{themes}
