% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{make_long}
\alias{make_long}
\title{Wide-to-Long}
\usage{
make_long(df_wide, s = NULL, x = "period")
}
\arguments{
\item{df_wide}{Wide dataframe (must have a period column)}

\item{s}{Named vector containing levels for the series column, in the desired order}

\item{x}{Column name for the period column}
}
\description{
Pivots wider, parses dates, renames and reorders levels for series.
}
\examples{
make_long(df_wide, c("Seasonal" = "BLDM.SFTZ1100A1S", "Actual" = "BLDM.SFTZ1100A1A"))
}
\keyword{clean}
\keyword{hud}
\keyword{long}
