% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{make_regional}
\alias{make_regional}
\title{Long-to-regional}
\usage{
make_regional(
  df_long,
  national = "National",
  r = "region",
  x = "period",
  y = "value"
)
}
\arguments{
\item{df_long}{Long dataframe}

\item{r}{Column name for the region column}

\item{x}{Column name for the period column}

\item{y}{Column name for the value column}
}
\description{
Using one series as a national baseline, create regional/national pairs for each region.
Region will go in region column, regional/national will go into a series column.
}
\examples{
make_regional(df_long)
}
\keyword{clean}
\keyword{hud}
\keyword{regional}
