% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_line}
\alias{plot_line}
\title{HUD Line Chart}
\usage{
plot_line(
  df_long,
  palette_type = "categorical",
  medium = "print",
  layout = "big",
  c = "series",
  x = "period",
  y = "value"
)
}
\arguments{
\item{df_long}{Dataframe produced by make_long() (e.g. series, period, value)}

\item{palette_type}{How to colour the series ("single", "double", "categorical")}

\item{medium}{"web" (dark background theme) or "print" (white background theme)}

\item{layout}{"big" or "small}

\item{c}{Name of column to use for colour}

\item{x}{Name of column to use for x-axis}

\item{y}{Name of column to use for y-axis}
}
\description{
Pre-made ggplot2 line charts using HUD styles.
}
\examples{
plot_long(df_long, palette_type = "double")
}
\keyword{charts}
\keyword{ggplot2}
\keyword{hud}
