% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_regional_lines}
\alias{plot_regional_lines}
\title{HUD Regional Line Charts}
\usage{
plot_regional_lines(
  df_reg,
  ncols = 1,
  r = "region",
  palette_type = "categorical",
  medium = "print",
  layout = "big",
  c = "series",
  x = "period",
  y = "value"
)
}
\arguments{
\item{ncols}{Number of columns for small multiples (facets) grid}

\item{r}{Name of column to use for region}

\item{palette_type}{How to colour the series ("single", "double", "categorical")}

\item{medium}{"web" (dark background theme) or "print" (white background theme)}

\item{layout}{"big" or "small}

\item{c}{Name of column to use for colour}

\item{x}{Name of column to use for x-axis}

\item{y}{Name of column to use for y-axis}

\item{df_long}{Dataframe produced by make_regional() (e.g. region, series, period, value)}
}
\description{
Pre-made ggplot2 small multiples (facets) line charts using HUD styles.
}
\examples{
plot_reg(df_reg, palette_type = "double")
}
\keyword{charts}
\keyword{ggplot2}
\keyword{hud}
