% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hud_theme}
\alias{hud_theme}
\title{HUD Themes}
\arguments{
\item{medium}{"print" (white background theme) or "web" (dark background theme)}

\item{layout}{"big" or "small}
}
\description{
ggplot2 theme generation function for HUD styles.
}
\examples{
hud_theme(medium = "web", layout = "big")
}
\keyword{ggplot2}
\keyword{hud}
\keyword{themes}
